+++ integration/rabbitmq/e2e/rpc.e2e-spec.ts
@@ -3,6 +3,8 @@ import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { AppModule } from '../src/app.module';
 
+const nonJsonRpcRoutingKey = 'non-json-rpc';
+
 describe('Rabbit RPC', () => {
   let app: INestApplication;
   let amqpConnection: AmqpConnection;
@@ -44,5 +46,48 @@ describe('Rabbit RPC', () => {
     });
 
     expect(response).toEqual({ transformed: { message: 42 } });
+  });
+
+  it('error reply RPC handler with non-JSON message should return an RPC error response', async () => {
+    const response = await amqpConnection.request({
+      exchange: 'exchange1',
+      routingKey: 'error-reply-rpc',
+      payload: Buffer.from('{a:'),
+    });
+
+    expect(response).toHaveProperty('message');
+    expect(response).toMatchObject({ status: 'error' });
+  });
+
+  it('non-JSON RPC handler with should receive a valid RPC response', async () => {
+    const response = await amqpConnection.request({
+      exchange: 'exchange1',
+      routingKey: nonJsonRpcRoutingKey,
+      payload: {
+        request: 'val',
+      },
+    });
+
+    expect(response).toEqual({ echo: { request: 'val' } });
+  });
+
+  it('non-JSON RPC handler with undefined message should receive a valid RPC response', async () => {
+    const response = await amqpConnection.request({
+      exchange: 'exchange1',
+      routingKey: nonJsonRpcRoutingKey,
+      payload: Buffer.alloc(0),
+    });
+
+    expect(response).toEqual({ echo: undefined });
+  });
+
+  it('non-JSON RPC handler with unparsable message should receive a valid RPC response', async () => {
+    const response = await amqpConnection.request({
+      exchange: 'exchange1',
+      routingKey: nonJsonRpcRoutingKey,
+      payload: Buffer.from('{a:'),
+    });
+
+    expect(response).toEqual({ echo: undefined });
   });
 });