+++ integration/rabbitmq/src/rpc/rpc.service.ts
@@ -1,6 +1,11 @@
-import { RabbitRPC } from '@golevelup/nestjs-rabbitmq';
+import {
+  MessageHandlerErrorBehavior,
+  RabbitRPC,
+} from '@golevelup/nestjs-rabbitmq';
 import { Injectable, UseInterceptors } from '@nestjs/common';
+import { RpcException } from '@nestjs/microservices';
 import { TransformInterceptor } from '../transform.interceptor';
+import { ReplyErrorCallback } from './reply.error.callback';
 
 @Injectable()
 export class RpcService {
@@ -24,6 +29,29 @@ export class RpcService {
   interceptedRpc() {
     return {
       message: 42,
+    };
+  }
+
+  @RabbitRPC({
+    routingKey: 'error-reply-rpc',
+    exchange: 'exchange1',
+    queue: 'error-reply-rpc',
+    errorBehavior: MessageHandlerErrorBehavior.ACK,
+    errorHandler: ReplyErrorCallback,
+  })
+  errorReplyRpc(message: object) {
+    throw new RpcException(message);
+  }
+
+  @RabbitRPC({
+    routingKey: 'non-json-rpc',
+    exchange: 'exchange1',
+    queue: 'non-json-rpc',
+    allowNonJsonMessages: true,
+  })
+  nonJsonRpc(nonJsonMessage: any) {
+    return {
+      echo: nonJsonMessage,
     };
   }
 }