+++ packages/hasura/src/hasura.event-handler.guard.ts
@@ -1,25 +1,20 @@
-import {
-  CanActivate,
-  ExecutionContext,
-  Inject,
-  Injectable,
-} from '@nestjs/common';
+import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { Request } from 'express';
 import { Observable } from 'rxjs';
-import { HASURA_MODULE_CONFIG } from './hasura.constants';
+import { InjectHasuraConfig } from './hasura.decorators';
 import { HasuraModuleConfig } from './hasura.interfaces';
 
 @Injectable()
 export class HasuraEventHandlerHeaderGuard implements CanActivate {
   private readonly apiSecret: string;
   constructor(
-    @Inject(HASURA_MODULE_CONFIG)
+    @InjectHasuraConfig()
     private readonly hasuraConfig: HasuraModuleConfig
   ) {
     this.apiSecret =
-      typeof hasuraConfig.secretFactory === 'function'
-        ? hasuraConfig.secretFactory()
-        : hasuraConfig.secretFactory;
+      typeof hasuraConfig.webhookConfig.secretFactory === 'function'
+        ? hasuraConfig.webhookConfig.secretFactory()
+        : hasuraConfig.webhookConfig.secretFactory;
   }
 
   canActivate(
@@ -27,7 +22,8 @@ export class HasuraEventHandlerHeaderGuard implements CanActivate {
   ): boolean | Promise<boolean> | Observable<boolean> {
     const request = context.switchToHttp().getRequest<Request>();
 
-    const secretRequestHeader = request.headers[this.hasuraConfig.secretHeader];
+    const secretRequestHeader =
+      request.headers[this.hasuraConfig.webhookConfig.secretHeader];
 
     return secretRequestHeader === this.apiSecret;
   }