+++ packages/rabbitmq/src/rabbitmq.interfaces.ts
@@ -1,5 +1,9 @@
 import * as amqpConnectionManager from 'amqp-connection-manager';
 import * as amqplib from 'amqplib';
+import {
+  MessageErrorHandler,
+  MessageHandlerErrorBehavior,
+} from './amqp/errorBehaviors';
 
 export interface RabbitMQExchangeConfig {
   name: string;
@@ -31,20 +35,25 @@ export interface QueueOptions {
   deadLetterRoutingKey?: string;
   maxLength?: number;
   maxPriority?: number;
-}
-
-export enum MessageHandlerErrorBehavior {
-  ACK,
-  NACK,
-  REQUEUE
+  bindQueueArguments?: any;
 }
 
 export interface MessageHandlerOptions {
-  exchange: string;
-  routingKey: string | string[];
+  exchange?: string;
+  routingKey?: string | string[];
   queue?: string;
   queueOptions?: QueueOptions;
+  /**
+   * @deprecated()
+   * Legacy error handling behaviors. This will be overridden if the errorHandler property is set
+   */
   errorBehavior?: MessageHandlerErrorBehavior;
+  /**
+   * A function that will be called if an error is thrown during processing of an incoming message
+   */
+  errorHandler?: MessageErrorHandler;
+  allowNonJsonMessages?: boolean;
+  createQueueIfNotExists?: boolean;
 }
 
 export interface ConnectionInitOptions {